# openapi/mcp-federation.openapi.yaml
openapi: 3.1.0
info:
  title: ChatGPT Gateway MCP Federation Control Plane
  version: 2.4.0
  description: Machine-readable MCP federation, presentation-profile, immutable projection,
    tenant-scoped hybrid catalog retrieval and roots control-plane contract.
paths:
...
  /api/mcp/credential-bindings/material:
    post:
      tags:
      - mcp-upstream
      summary: Create Material Binding
      operationId: create_material_binding_api_mcp_credential_bindings_material_post
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 160
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpCredentialMaterialCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpCredentialBindingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/mcp/credential-bindings/{binding_id}/rotate:
    post:
      tags:
      - mcp-upstream
      summary: Rotate Material Binding
      operationId: rotate_material_binding_api_mcp_credential_bindings__binding_id__rotate_post
      parameters:
      - name: binding_id
        in: path
        required: true
        schema:
          type: string
          title: Binding Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 160
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpCredentialMaterialRotate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpCredentialBindingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/mcp/credential-bindings/{binding_id}/revoke:
    post:
      tags:
      - mcp-upstream
      summary: Revoke Material Binding
      operationId: revoke_material_binding_api_mcp_credential_bindings__binding_id__revoke_post
      parameters:
      - name: binding_id
        in: path
        required: true
        schema:
          type: string
          title: Binding Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpCredentialCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpCredentialBindingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/mcp/servers/{server_id}/oauth/discover:
    post:
      tags:
      - mcp-upstream
      summary: Oauth Discover
      operationId: oauth_discover_api_mcp_servers__server_id__oauth_discover_post
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpOAuthDiscoveryRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpOAuthDiscoveryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/mcp/servers/{server_id}/oauth/start:
    post:
      tags:
      - mcp-upstream
      summary: Oauth Start
      operationId: oauth_start_api_mcp_servers__server_id__oauth_start_post
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 160
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpOAuthAuthorizationStart'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpOAuthAuthorizationStarted'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/mcp/oauth/complete:
    post:
      tags:
      - mcp-upstream
      summary: Oauth Complete
      operationId: oauth_complete_api_mcp_oauth_complete_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpOAuthAuthorizationComplete'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpCredentialBindingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/mcp/servers/{server_id}/runtime-call:
    post:
      tags:
      - mcp-upstream
      summary: Runtime Call
      operationId: runtime_call_api_mcp_servers__server_id__runtime_call_post
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 160
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpUpstreamCallInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpUpstreamCallOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    McpCredentialBindingOut:
      properties:
        id:
          type: string
          title: Id
        owner_subject:
          type: string
          title: Owner Subject
        binding_type:
          type: string
          title: Binding Type
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        audience:
          anyOf:
          - type: string
          - type: 'null'
          title: Audience
        scopes:
          items:
            type: string
          type: array
          title: Scopes
        status:
          type: string
          title: Status
        version:
          type: integer
          title: Version
        meta:
          additionalProperties: true
          type: object
          title: Meta
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        rotated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Rotated At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
      type: object
      required:
      - id
      - owner_subject
      - binding_type
      - scopes
      - status
      - version
      - meta
      - created_at
      - updated_at
      title: McpCredentialBindingOut
    McpCredentialCommand:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
      additionalProperties: false
      type: object
      required:
      - expected_version
      title: McpCredentialCommand
    McpCredentialMaterialCreate:
      properties:
        binding_type:
          type: string
          enum:
          - oauth
          - service_account
          title: Binding Type
        provider:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Provider
        audience:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Audience
        issuer:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Issuer
        scopes:
          items:
            type: string
          type: array
          maxItems: 100
          title: Scopes
        mode:
          type: string
          enum:
          - bearer
          - header
          - oauth
          title: Mode
        access_token:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Access Token
        refresh_token:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Refresh Token
        token_endpoint:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Token Endpoint
        client_id:
          anyOf:
          - type: string
            maxLength: 512
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Client Secret
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        header_name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Header Name
        header_value:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Header Value
      additionalProperties: false
      type: object
      required:
      - binding_type
      - mode
      title: McpCredentialMaterialCreate
    McpCredentialMaterialRotate:
      properties:
        binding_type:
          type: string
          enum:
          - oauth
          - service_account
          title: Binding Type
        provider:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Provider
        audience:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Audience
        issuer:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Issuer
        scopes:
          items:
            type: string
          type: array
          maxItems: 100
          title: Scopes
        mode:
          type: string
          enum:
          - bearer
          - header
          - oauth
          title: Mode
        access_token:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Access Token
        refresh_token:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Refresh Token
        token_endpoint:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Token Endpoint
        client_id:
          anyOf:
          - type: string
            maxLength: 512
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Client Secret
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        header_name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Header Name
        header_value:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Header Value
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
      additionalProperties: false
      type: object
      required:
      - binding_type
      - mode
      - expected_version
      title: McpCredentialMaterialRotate
...
    McpOAuthAuthorizationComplete:
      properties:
        state:
          type: string
          maxLength: 512
          minLength: 20
          title: State
        code:
          type: string
          format: password
          title: Code
          writeOnly: true
        iss:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Iss
      additionalProperties: false
      type: object
      required:
      - state
      - code
      title: McpOAuthAuthorizationComplete
    McpOAuthAuthorizationStart:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
        discovery_snapshot_id:
          anyOf:
          - type: string
            maxLength: 36
          - type: 'null'
          title: Discovery Snapshot Id
        authorization_server:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Authorization Server
        authorization_endpoint:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Authorization Endpoint
        token_endpoint:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Token Endpoint
        client_id:
          type: string
          maxLength: 512
          minLength: 1
          title: Client Id
        client_secret:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Client Secret
        redirect_uri:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Redirect Uri
        scopes:
          items:
            type: string
          type: array
          maxItems: 100
          title: Scopes
        audience:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Audience
        extra_authorization_parameters:
          additionalProperties:
            type: string
          type: object
          title: Extra Authorization Parameters
      additionalProperties: false
      type: object
      required:
      - expected_version
      - client_id
      - redirect_uri
      title: McpOAuthAuthorizationStart
    McpOAuthAuthorizationStarted:
      properties:
        server_id:
          type: string
          title: Server Id
        binding_id:
          type: string
          title: Binding Id
        authorization_url:
          type: string
          title: Authorization Url
        state:
          type: string
          title: State
        expires_at:
          type: string
          format: date-time
          title: Expires At
      type: object
      required:
      - server_id
      - binding_id
      - authorization_url
      - state
      - expires_at
      title: McpOAuthAuthorizationStarted
    McpOAuthDiscoveryOut:
      properties:
        id:
          type: string
          title: Id
        server_id:
          type: string
          title: Server Id
        resource:
          type: string
          title: Resource
        resource_metadata_url:
          type: string
          title: Resource Metadata Url
        authorization_server:
          type: string
          title: Authorization Server
        authorization_server_metadata_url:
          type: string
          title: Authorization Server Metadata Url
        discovery_mechanism:
          type: string
          title: Discovery Mechanism
        authorization_endpoint:
          type: string
          title: Authorization Endpoint
        token_endpoint:
          type: string
          title: Token Endpoint
        registration_endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Registration Endpoint
        requested_scopes:
          items:
            type: string
          type: array
          title: Requested Scopes
        proposed_scopes:
          items:
            type: string
          type: array
          title: Proposed Scopes
        metadata_hash:
          type: string
          title: Metadata Hash
        expires_at:
          type: string
          format: date-time
          title: Expires At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - server_id
      - resource
      - resource_metadata_url
      - authorization_server
      - authorization_server_metadata_url
      - discovery_mechanism
      - authorization_endpoint
      - token_endpoint
      - requested_scopes
      - proposed_scopes
      - metadata_hash
      - expires_at
      - created_at
      title: McpOAuthDiscoveryOut
    McpOAuthDiscoveryRequest:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
        requested_scopes:
          items:
            type: string
          type: array
          maxItems: 100
          title: Requested Scopes
        authorization_server:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Authorization Server
      additionalProperties: false
      type: object
      required:
      - expected_version
      title: McpOAuthDiscoveryRequest
...
    McpUpstreamCallInput:
      properties:
        revision_id:
          type: string
          maxLength: 36
          minLength: 1
          title: Revision Id
        arguments:
          additionalProperties: true
          type: object
          title: Arguments
        timeout_seconds:
          type: number
          maximum: 300.0
          minimum: 0.1
          title: Timeout Seconds
          default: 30.0
      additionalProperties: false
      type: object
      required:
      - revision_id
      title: McpUpstreamCallInput
    McpUpstreamCallOut:
      properties:
        revision_id:
          type: string
          title: Revision Id
        schema_hash:
          type: string
          title: Schema Hash
        result:
          additionalProperties: true
          type: object
          title: Result
        truncated:
          type: boolean
          title: Truncated
        serialized_bytes:
          type: integer
          title: Serialized Bytes
      type: object
      required:
      - revision_id
      - schema_hash
      - result
      - truncated
      - serialized_bytes
      title: McpUpstreamCallOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError